Move auth and user-agent out
