+++ codex-rs/backend-client/Cargo.toml
@@ -15,7 +15,8 @@ serde_json = "1"
 reqwest = { version = "0.12", default-features = false, features = ["json", "rustls-tls"] }
 codex-backend-openapi-models = { path = "../codex-backend-openapi-models" }
 codex-protocol = { workspace = true }
-codex-core = { workspace = true }
+codex-auth = { workspace = true }
+codex-user-agent = { workspace = true }
 
 [dev-dependencies]
 pretty_assertions = "1"
