+++ codex-rs/core/src/util.rs
@@ -3,7 +3,6 @@ use std::path::PathBuf;
 use std::time::Duration;
 
 use rand::Rng;
-use tracing::debug;
 use tracing::error;
 
 const INITIAL_DELAY_MS: u64 = 200;
@@ -49,21 +48,6 @@ pub(crate) fn error_or_panic(message: impl std::string::ToString) {
     }
 }
 
-pub(crate) fn try_parse_error_message(text: &str) -> String {
-    debug!("Parsing server error response: {}", text);
-    let json = serde_json::from_str::<serde_json::Value>(text).unwrap_or_default();
-    if let Some(error) = json.get("error")
-        && let Some(message) = error.get("message")
-        && let Some(message_str) = message.as_str()
-    {
-        return message_str.to_string();
-    }
-    if text.is_empty() {
-        return "Unknown error".to_string();
-    }
-    text.to_string()
-}
-
 pub fn resolve_path(base: &Path, path: &PathBuf) -> PathBuf {
     if path.is_absolute() {
         path.clone()
@@ -74,37 +58,11 @@ pub fn resolve_path(base: &Path, path: &PathBuf) -> PathBuf {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-
-    #[test]
-    fn test_try_parse_error_message() {
-        let text = r#"{
-  "error": {
-    "message": "Your refresh token has already been used to generate a new access token. Please try signing in again.",
-    "type": "invalid_request_error",
-    "param": null,
-    "code": "refresh_token_reused"
-  }
-}"#;
-        let message = try_parse_error_message(text);
-        assert_eq!(
-            message,
-            "Your refresh token has already been used to generate a new access token. Please try signing in again."
-        );
-    }
-
-    #[test]
-    fn test_try_parse_error_message_no_error() {
-        let text = r#"{"message": "test"}"#;
-        let message = try_parse_error_message(text);
-        assert_eq!(message, r#"{"message": "test"}"#);
-    }
-
     #[test]
     fn feedback_tags_macro_compiles() {
         #[derive(Debug)]
         struct OnlyDebug;
 
-        feedback_tags!(model = "gpt-5", cached = true, debug_only = OnlyDebug);
+        crate::feedback_tags!(model = "gpt-5", cached = true, debug_only = OnlyDebug);
     }
 }
