+++ codex-rs/user-agent/BUILD.bazel
@@ -1,0 +1,6 @@
+load("//:defs.bzl", "codex_rust_crate")
+
+codex_rust_crate(
+    name = "user-agent",
+    crate_name = "codex_user_agent",
+)
